move more tests to mls_group/tests_and_kats
