+++ openmls/src/group/mls_group/tests_and_kats/mod.rs
@@ -2,7 +2,9 @@
 
 mod kat_passive_client;
 mod kat_welcome;
+mod test_core_group;
 mod test_create_commit_params;
 mod test_external_init;
 mod test_mls_group;
 mod test_past_secrets;
+pub(crate) mod utils;
